requirements.txt
backports.functools-lru-cache==1.5
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Django==1.11.16
django-crispy-forms==1.7.2
django-reverse-admin==2.3.0
djangorestframework==3.9.0
get==1.0.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
post==1.0.2
public==2.0.1
pytz==2018.7
query-string==1.0.2
request==1.0.2
requests==2.21.0
soupsieve==1.7.3
urllib3==1.24.1

